# internal/product/handler.go
// HTTP handlers for products. Public read routes (list/get/export) and
// admin-only write routes, all returning the shared response envelope.
package product

import (
	"strconv"
	"strings"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"

	"inventory/internal/shared/audit"
	"inventory/internal/shared/dbutil"
	sharederr "inventory/internal/shared/errors"
	"inventory/internal/shared/export"
	"inventory/internal/shared/response"
	"inventory/internal/shared/validator"
)

// Handler exposes product routes.
type Handler struct {
	svc   *Service
	val   *validator.Validator
	audit audit.Recorder
}

// NewHandler wires the service and validator.
func NewHandler(svc *Service, val *validator.Validator) *Handler {
	return &Handler{svc: svc, val: val, audit: audit.Nop{}}
}

// SetAudit wires an audit recorder (nil-safe; Nop by default).
func (h *Handler) SetAudit(r audit.Recorder) {
	if r != nil {
		h.audit = r
	}
}

// record captures an audit event for a mutation, attaching the acting user
// and request IP. Details are nil-safe.
func (h *Handler) record(c *gin.Context, action, entityID string, details gin.H) {
	eid := entityID
	h.audit.Record(audit.EntryFromContext(c, audit.Entry{
		Action:     action,
		EntityType: "product",
		EntityID:   &eid,
		Details:    details,
	}))
}

type listProductsRequest struct {
	Q          string   `form:"q"`
	CategoryID string   `form:"category_id"`
	MinPrice   *float64 `form:"min_price"`
	MaxPrice   *float64 `form:"max_price"`
	LowStock   bool     `form:"low_stock"`
	Archived   string   `form:"is_archived"`
	Sort       string   `form:"sort"`
	Page       int      `form:"page"`
	PerPage    int      `form:"per_page"`
}

type createProductRequest struct {
	Name              string  `json:"name" validate:"required,min=2"`
	SKU               string  `json:"sku" validate:"required,min=2"`
	Description       *string `json:"description"`
	Price             float64 `json:"price"`
	CategoryID        string  `json:"category_id" validate:"required"`
	LowStockThreshold int     `json:"low_stock_threshold"`
	IsArchived        bool    `json:"is_archived"`
}

type updateProductRequest struct {
	Name              *string  `json:"name"`
	SKU               *string  `json:"sku"`
	Description       *string  `json:"description"`
	Price             *float64 `json:"price"`
	CategoryID        *string  `json:"category_id"`
	LowStockThreshold *int     `json:"low_stock_threshold"`
	IsArchived        *bool    `json:"is_archived"`
}

// productEnvelope is the JSON shape returned by the API. It includes the
// category name alongside the id for client convenience.
type productEnvelope struct {
	ID                uuid.UUID `json:"id"`
	Name              string    `json:"name"`
	SKU               string    `json:"sku"`
	Description       *string   `json:"description,omitempty"`
	Price             float64   `json:"price"`
	CategoryID        uuid.UUID `json:"category_id"`
	CategoryName      string    `json:"category_name"`
	LowStockThreshold int       `json:"low_stock_threshold"`
	IsArchived        bool      `json:"is_archived"`
	CreatedAt         string    `json:"created_at"`
	UpdatedAt         string    `json:"updated_at"`
}

func productResponse(p *Product) productEnvelope {
	return productEnvelope{
		ID:                p.ID,
		Name:              p.Name,
		SKU:               p.SKU,
		Description:       p.Description,
		Price:             p.Price,
		CategoryID:        p.CategoryID,
		CategoryName:      p.Category.Name,
		LowStockThreshold: p.LowStockThreshold,
		IsArchived:        p.IsArchived,
		CreatedAt:         p.CreatedAt.UTC().Format("2006-01-02T15:04:05Z"),
		UpdatedAt:         p.UpdatedAt.UTC().Format("2006-01-02T15:04:05Z"),
	}
}

func parseCategoryID(raw string) (uuid.UUID, bool) {
	if strings.TrimSpace(raw) == "" {
		return uuid.Nil, false
	}
	id, err := uuid.Parse(raw)
	return id, err == nil
}

func toBoolPtr(v string) *bool {
	if v == "" {
		return nil
	}
	b := v == "true" || v == "1"
	return &b
}

// List handles GET /products.
// @Tags products
// @Summary List products
// @Accept json
// @Produce json
// @Param q query string false "Search by name/SKU"
// @Param category_id query string false "Filter by category UUID"
// @Param min_price query number false "Minimum price"
// @Param max_price query number false "Maximum price"
// @Param low_stock query boolean false "Only low-stock items"
// @Param is_archived query string false "Archive filter" Enums(true, false)
// @Param sort query string false "Sort field/direction"
// @Param page query int false "Page number" default(1)
// @Param per_page query int false "Items per page" default(20)
// @Success 200 {object} response.Body
// @Failure 400 {object} response.Body
// @Router /products [get]
func (h *Handler) List(c *gin.Context) {
	var req listProductsRequest
	if err := c.ShouldBindQuery(&req); err != nil {
		response.Error(c, sharederr.ErrValidation)
		return
	}

	var catID uuid.UUID
	if req.CategoryID != "" {
		id, ok := parseCategoryID(req.CategoryID)
		if !ok {
			response.Error(c, sharederr.ErrValidation)
			return
		}
		catID = id
	}

	prods, total, err := h.svc.List(c.Request.Context(), ListQuery{
		Q:          req.Q,
		CategoryID: catID,
		MinPrice:   req.MinPrice,
		MaxPrice:   req.MaxPrice,
		LowStock:   req.LowStock,
		IsArchived: toBoolPtr(req.Archived),
		Sort:       req.Sort,
		Page:       req.Page,
		PerPage:    req.PerPage,
	})
	if err != nil {
		response.Error(c, err)
		return
	}

	items := make([]productEnvelope, 0, len(prods))
	for _, p := range prods {
		items = append(items, productResponse(p))
	}

	page, perPage := dbutil.NormalizePage(req.Page, req.PerPage)
	totalPages := 0
	if perPage > 0 {
		totalPages = int((total + int64(perPage) - 1) / int64(perPage))
	}
	response.Paginated(c, items, &response.Pagination{
		Page:       page,
		PerPage:    perPage,
		Total:      total,
		TotalPages: totalPages,
	})
}

// Get handles GET /products/:id.
// @Tags products
// @Summary Get a product
// @Accept json
// @Produce json
// @Param id path string true "Product ID"
// @Success 200 {object} response.Body
// @Failure 400 {object} response.Body
// @Failure 404 {object} response.Body
// @Router /products/{id} [get]
func (h *Handler) Get(c *gin.Context) {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		response.Error(c, sharederr.ErrValidation)
		return
	}
	p, err := h.svc.Get(c.Request.Context(), id)
	if err != nil {
		response.Error(c, err)
		return
	}
	response.OK(c, productResponse(p))
}

// Create handles POST /products.
// @Tags products
// @Summary Create a product
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param body body createProductRequest true "Product payload"
// @Success 201 {object} response.Body
// @Failure 400 {object} response.Body
// @Failure 401 {object} response.Body
// @Failure 403 {object} response.Body
// @Failure 404 {object} response.Body
// @Failure 409 {object} response.Body
// @Router /products [post]
func (h *Handler) Create(c *gin.Context) {
	var req createProductRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		response.Error(c, sharederr.ErrValidation)
		return
	}
	if err := h.val.Validate(req); err != nil {
		response.Error(c, sharederr.ErrValidation)
		return
	}
	catID, ok := parseCategoryID(req.CategoryID)
	if !ok {
		response.Error(c, sharederr.ErrValidation)
		return
	}

	p, err := h.svc.Create(
		c.Request.Context(), req.Name, req.SKU, req.Description,
		req.Price, catID, req.LowStockThreshold, req.IsArchived,
	)
	if err != nil {
		response.Error(c, err)
		return
	}
	h.record(c, "CREATE", p.ID.String(), gin.H{"name": p.Name, "sku": p.SKU})
	response.Created(c, productResponse(p))
}

// Update handles PUT /products/:id.
// @Tags products
// @Summary Update a product
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path string true "Product ID"
// @Param body body updateProductRequest true "Product update payload"
// @Success 200 {object} response.Body
// @Failure 400 {object} response.Body
// @Failure 401 {object} response.Body
// @Failure 403 {object} response.Body
// @Failure 404 {object} response.Body
// @Failure 409 {object} response.Body
// @Router /products/{id} [put]
func (h *Handler) Update(c *gin.Context) {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		response.Error(c, sharederr.ErrValidation)
		return
	}

	var req updateProductRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		response.Error(c, sharederr.ErrValidation)
		return
	}
	if req.Name == nil && req.SKU == nil && req.Description == nil && req.Price == nil &&
		req.CategoryID == nil && req.LowStockThreshold == nil && req.IsArchived == nil {
		response.Error(c, sharederr.ErrValidation)
		return
	}

	var catID *uuid.UUID
	if req.CategoryID != nil {
		cid, ok := parseCategoryID(*req.CategoryID)
		if !ok {
			response.Error(c, sharederr.ErrValidation)
			return
		}
		catID = &cid
	}

	p, err := h.svc.Update(c.Request.Context(), id, UpdateParams{
		Name:              req.Name,
		SKU:               req.SKU,
		Description:       req.Description,
		Price:             req.Price,
		CategoryID:        catID,
		LowStockThreshold: req.LowStockThreshold,
		IsArchived:        req.IsArchived,
	})
	if err != nil {
		response.Error(c, err)
		return
	}
	h.record(c, "UPDATE", p.ID.String(), gin.H{"name": p.Name, "sku": p.SKU})
	response.OK(c, productResponse(p))
}

// Delete handles DELETE /products/:id.
// @Tags products
// @Summary Delete a product
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path string true "Product ID"
// @Success 200 {object} response.Body
// @Failure 400 {object} response.Body
// @Failure 401 {object} response.Body
// @Failure 403 {object} response.Body
// @Failure 404 {object} response.Body
// @Failure 409 {object} response.Body
// @Router /products/{id} [delete]
func (h *Handler) Delete(c *gin.Context) {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		response.Error(c, sharederr.ErrValidation)
		return
	}
	if err := h.svc.Delete(c.Request.Context(), id); err != nil {
		response.Error(c, err)
		return
	}
	h.record(c, "DELETE", id.String(), gin.H{"archived": true})
	response.Message(c, "product archived")
}

// Export handles GET /products/export — CSV download of all products.
// @Tags products
// @Summary Export products as CSV
// @Accept json
// @Produce text/csv
// @Success 200
// @Router /products/export [get]
func (h *Handler) Export(c *gin.Context) {
	// The list repository caps per_page (≤100), so page through the whole
	// catalog instead of requesting an unclamped page.
	var prods []*Product
	for page := 1; ; page++ {
		pageProds, total, err := h.svc.List(c.Request.Context(), ListQuery{Page: page, PerPage: 100})
		if err != nil {
			response.Error(c, err)
			return
		}
		prods = append(prods, pageProds...)
		if len(pageProds) == 0 || len(prods) >= int(total) {
			break
		}
	}

	rows := make([][]string, 0, len(prods))
	for _, p := range prods {
		desc := ""
		if p.Description != nil {
			desc = *p.Description
		}
		rows = append(rows, []string{
			p.ID.String(),
			p.Name,
			p.SKU,
			strconv.FormatFloat(p.Price, 'f', 2, 64),
			p.Category.Name,
			strconv.Itoa(p.LowStockThreshold),
			strconv.FormatBool(p.IsArchived),
			desc,
			p.CreatedAt.UTC().Format("2006-01-02T15:04:05Z"),
		})
	}

	export.SetAttachment(c, "products")
	headers := []string{
		"id", "name", "sku", "price", "category_name",
		"low_stock_threshold", "is_archived", "description", "created_at",
	}
	if err := export.WriteCSV(c.Writer, headers, rows); err != nil {
		return
	}
}